requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
pydantic==1.10.12
requests==2.31.0
SQLAlchemy==1.4.49
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
